% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addTokenId.R
\name{addTokenId}
\alias{addTokenId}
\title{Function to add a unique identifier for each token to a concordance}
\usage{
addTokenId(concordance)
}
\arguments{
\item{concordance}{a concordance generated with makeConcordance}
}
\value{
a concordance data.frame, with a tokenId column added
}
\description{
Function to add a unique identifier for each token to a concordance
}
