% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cleanupText.R
\name{cleanupText}
\alias{cleanupText}
\title{Function to clean up common encoding issues}
\usage{
cleanupText(characterVector)
}
\arguments{
\item{character.vector}{Character vector or string you want to clean up}
}
\value{
cleaned character vector
}
\description{
Function to clean up common encoding issues
}
