% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/corpusStatistics.R
\name{corpusStatistics}
\alias{corpusStatistics}
\title{Function to generate summary statistics for a Twitter corpus}
\usage{
corpusStatistics(corpus_directory)
}
\arguments{
\item{corpus_directory}{path to the directory where you keep the corpus CSV files. With trailing slash}
}
\value{
a data.frame that enumerates per file the number of tweets, the number of users, and the number of words
}
\description{
Function to generate summary statistics for a Twitter corpus
}
