% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geoNames.R
\name{GeoNames}
\alias{GeoNames}
\title{GeoNames: function to create a geonames database file from geonames.org export}
\usage{
GeoNames(GeoNames_file, AlternateNames_file, output_file = paste0(getwd(),
  "/GeoNames_outputfile.csv"))
}
\arguments{
\item{GeoNames_file}{download/unzip this file http://download.geonames.org/export/dump/allCountries.zip and specify its path (e.g. GeoNames_file="~/Downloads/allCountries.txt")}

\item{AlternateNames_file}{download/unzip this file http://download.geonames.org/export/dump/alternateNames.zip and specify its path}

\item{output_file}{Where should the output go? Defaults to the current working directory/GeoNames_outputfile.csv}
}
\value{
A message signaling completion. The actual result is not returned but rather written to a CSV file
}
\description{
This function will create a geonames database file which includes alternative spellings of many placenames. The function reads/munges a lot of data, so give it some time.
}
