% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logmessage.R
\name{logmessage}
\alias{logmessage}
\title{Simple function to generate logs with timestamps}
\usage{
logmessage(string, verbose = T)
}
\arguments{
\item{string}{The message}

\item{verbose}{Verbosity: if TRUE, the function will print the message, if FALSE the function will do nothing at all}
}
\value{
None. The message with an added timestamp will be printed to the screen
}
\description{
Simple function to generate logs with timestamps
}
