% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeConcordance.R
\name{makeConcordance}
\alias{makeConcordance}
\title{Function to create a concordance from a filtered corpus data.frame}
\usage{
makeConcordance(filtered_corpus, pattern = NULL, column = c("tagged",
  "text"))
}
\arguments{
\item{filtered_corpus}{a data.frame generated by searchCorpus or by calling command-line grep on the corpus files}

\item{pattern}{a regex pattern. If you want to match full-word strings, just pad your string with the word boundary symbol \code{\\b}. E.g. \code{\\b(brother)\\b}}

\item{column}{column to match the pattern against. The column 'text' contains the untagged text, the column 'text' contains the dependency-tagged text}
}
\value{
a data.frame with a conconcordance for your pattern added to the input data.frame
}
\description{
This function will create a concordance for a string or regex pattern
}
\examples{
this function can be combined with searchCorpus in a corpus-processing pipeline. For example: \\code{concordance<-searchCorpus() \%>\% 
makeConcordance() }
}
