% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SpanishNames.R
\name{SpanishNames}
\alias{SpanishNames}
\title{Function to create a babynames statistics database for Spanish names}
\usage{
SpanishNames(verbose = T)
}
\arguments{
\item{verbose}{if TRUE generate verbose output}
}
\value{
a babynames statistics database, listing for each name the most likely gender and year of birth
}
\description{
Function to create a babynames statistics database for Spanish names
}
