% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/syntaxNetR.R
\name{syntaxNetR}
\alias{syntaxNetR}
\title{\code{syntaxNetR} a basic R binding for Google SyntaxNet Parsey Universal}
\usage{
syntaxNetR(tokenVector, syntaxNetPath = NULL,
  language = c("Ancient_Greek-PROIEL", "Ancient_Greek", "Arabic", "Basque",
  "Bulgarian", "Catalan", "Chinese", "Croatian", "Czech-CAC", "Czech-CLTT",
  "Czech", "Danish", "Dutch-LassySmall", "Dutch", "English-LinES", "English",
  "Estonian", "Finnish-FTB", "Finnish", "French", "Galician", "German",
  "Gothic", "Greek", "Hebrew", "Hindi", "Hungarian", "Indonesian", "Irish",
  "Italian", "Kazakh", "Latin-ITTB", "Latin-PROIEL", "Latin", "Latvian",
  "Norwegian", "Old_Church_Slavonic", "Persian", "Polish", "Portuguese-BR",     
  "Portuguese", "Romanian", "Russian-SynTagRus", "Russian", "Slovenian-SST",
  "Slovenian", "Spanish-AnCora", "Spanish", "Swedish-LinES", "Swedish", "Tamil",
  "Turkish"), SIMPLIFY = TRUE, verbose = TRUE)
}
\arguments{
\item{tokenVector}{a character vector formatted in UTF-8 or ASCII. It will almost certainly not work with windows/ISO 8591-1/latin1 encoding.}

\item{syntaxNetPath}{path to your local syntaxNet installation, with trailing slashes (i.e., the path has to end with a forward slash)}

\item{language}{name of the language  model provided by Google, must be one of pre-set options}

\item{SIMPLIFY}{if set to TRUE (default), the function will simplify the list output to text with slashtags (e.g., _noun_sg_nsubj)}

\item{verbose}{if set to TRUE (default) the function will output log messages allowing you to track its progress}
}
\value{
a parsed character vector (if SIMPLIFY is set to TRUE) or a list representation of the parsed character vector
}
\description{
Basic R binding to use an adapted version of the example python code provided by Google to tag R character vectors. You will need a Mac/Linux computer with an installation of syntaxNet to use this function. Follow the steps described here:https://github.com/tensorflow/models/tree/master/syntaxnet.
}
